Splitter code
Start is the image produced by pretrained model. All is image produced by using entire dataset, best is using the 10% most upvoted images and worst is using the 10% least upvoted images. These images are taken from the models with the least FID value and can be split up running the main.py in the Splitter folder.